Half fix blocking not always kicking.
Sometimes the blocker doesn't work in disconnecting you, and you stay connected, this is due to it being ran as soon you join a room.

You now get kicked once any admin event is ran, or a player joins. This ruins the game in publics and if you were annoying an admin in a private stops you from doing that.